feat(taskee skillset): refactor controller
 - move controller logic from task to skillset
 - add toast notification on duplicate entry

[Finishes #117835155]
