+++ app/helpers/skillsets_helper.rb
@@ -1,0 +1,2 @@
+module SkillsetsHelper
+end
