+++ app/views/partials/_sidebar_dashboard.html.erb
@@ -4,10 +4,10 @@
     <%= link_to raw("<li class='side-nav-button'>Edit Profile</li>"), profile_path %><hr>
     <%= link_to raw("<li class='side-nav-button'>Task Management</li>"), dashboard_path %><hr>
     <%= link_to raw("<li class='side-nav-button'>Account Upgrade</li>"), dashboard_path %><hr>
-    <% if current_user.user_type == "taskee" %>
-      <%= link_to raw("<li class='side-nav-button'>My Skillset</li>"), tasks_path %> <hr>
+    <% if current_user.taskee? %>
+      <%= link_to raw("<li class='side-nav-button'>My Skillset</li>"), skillsets_path %> <hr>
     <% end %>
-    <% if current_user.user_type == "tasker"%>
+    <% if current_user.tasker? %>
       <%= link_to raw("<li class='side-nav-button'>Map Search</li>"), map_search_path %><hr>
     <% end %>
   </ul>
