+++ app/views/skillsets/create.js.erb
@@ -1,0 +1,7 @@
+<% if @skillset.was_created %>
+  $('ul.collection').append("<%= j(render partial: 'skillset', object: @skillset.task) %>");
+  $('input#task_name').val('').removeClass('invalid');
+<% else %>
+  $('input#task_name').val('').addClass('invalid');
+  Materialize.toast("<%= @skillset.task.name %> already exists.", 4000);
+<% end %>
