+++ app/views/skillsets/destroy.js.erb
