+++ app/views/skillsets/index.html.erb
@@ -4,12 +4,12 @@
     <div class="col m10 offset-m1">
       <div class="row">
         <div class="col m5 offset-m1">
-          <% if @tasks.empty? %>
+          <% if @skillsets.empty? %>
             <p>You don't have any skillsets</p>
           <% else %>
             <ul class="collection with-header">
               <li class="collection-header"><h4>My Skillsets</h4></li>
-              <%= render @tasks %>
+              <%= render partial: 'skillset', collection: @skillsets %>
             </ul>
           <% end %>
         </div>
