--- app/views/tasks/_task.html.erb
@@ -1,7 +1,0 @@
-<li class="collection-item">
-  <%= div_for task do %>
-    <%= task.name %><%= link_to task, method: :delete, remote: :true, class: "secondary-content" do %>
-      <i class="material-icons">delete</i>
-    <% end %>
-  <% end %>
-</li>
