--- app/views/tasks/create.js.erb
@@ -1,6 +1,0 @@
-<% if @skillset.was_created %>
-  $('ul.collection').append("<%= j(render @task) %>");
-  $('input#task_name').val('').removeClass('invalid');
-<% else %>
-  $('input#task_name').val('').addClass('invalid');
-<% end %>
